Bug fix: prevent underflow during encryption.
The line:

  while (i < sz - AES_BLOCK_SIZE)

... had two problems. First, if sz < AES_BLOCK_SIZE, a buffer overflow occurs.
This causes a crash when encrypting small files (< 16 bytes). Rewriting the
exprission as (i + AES_BLOCK_SIZE < sz) fixes this.

The less serious problem is that the comparison was strict. But if sz ==
i + AES_BLOCK_SIZE, we can process another whole block. So we should write
<= instead of <. This is not a serious bug because the while-loop is followed
by some code to encrypt the remainder of the buffer byte-by-byte, so the worst
thing that could happen is that the final block is decoded in a slightly less
efficient way.